+++ src/yup/animations.ts
@@ -3,26 +3,14 @@ import { animate, AnimationMetadata, state, style, transition, trigger } from '@
 // Component transition animations
 export const mask: AnimationMetadata =
 trigger('mask', [
-    state('on', style({
-        opacity: 0.6
-    })),
-    state('init', style({
-        opacity: 0
-    })),
-    state('off', style({
-        opacity: 0
-    })),
-    transition('init => on', [
-        style({opacity: 0}), animate('0.3s linear')
-    ]),
-    transition('on => off', [
-        animate('0.3s linear',
-        style({opacity: 0}))
-    ])
+    state('enter', style({opacity: 0.7})),
+    state('void', style({opacity: 0})),
+    state('exit', style({opacity: 0})),
+    transition('* => *', animate('0.3s linear'))
 ]);
 export const dialog: AnimationMetadata =
 trigger('dialog', [
-    state('on', style({
+    state('enter', style({
         'opacity': '1',
         '-webkit-transform': 'translate(-50%, -50%) scale(1)',
         '-moz-transform': 'translate(-50%, -50%) scale(1)',
@@ -30,7 +18,7 @@ trigger('dialog', [
         '-o-transform': 'translate(-50%, -50%) scale(1)',
         'transform': 'translate(-50%, -50%) scale(1)'
     })),
-    state('init', style({
+    state('void', style({
         'opacity': '0',
         '-webkit-transform': 'translate(-50%, -50%) scale(1.2)',
         '-moz-transform': 'translate(-50%, -50%) scale(1.2)',
@@ -38,7 +26,7 @@ trigger('dialog', [
         '-o-transform': 'translate(-50%, -50%) scale(1.2)',
         'transform': 'translate(-50%, -50%) scale(1.2)'
     })),
-    state('off', style({
+    state('exit', style({
         'opacity': '0',
         '-webkit-transform': 'translate(-50%, -50%) scale(0.8)',
         '-moz-transform': 'translate(-50%, -50%) scale(0.8)',
@@ -46,33 +34,43 @@ trigger('dialog', [
         '-o-transform': 'translate(-50%, -50%) scale(0.8)',
         'transform': 'translate(-50%, -50%) scale(0.8)'
     })),
-    transition('init => on', [
-        style({
-            'opacity': '0',
-            '-webkit-transform': 'translate(-50%, -50%) scale(1.2)',
-            '-moz-transform': 'translate(-50%, -50%) scale(1.2)',
-            '-ms-transform': 'translate(-50%, -50%) scale(1.2)',
-            '-o-transform': 'translate(-50%, -50%) scale(1.2)',
-            'transform': 'translate(-50%, -50%) scale(1.2)'
-        }),
+    transition('* => *', [
         animate('0.3s linear')
-    ]),
-    transition('on => off', [
-        animate('0.3s linear',
-        style({
-            'opacity': '0',
-            '-webkit-transform': 'translate(-50%, -50%) scale(0.8)',
-            '-moz-transform': 'translate(-50%, -50%) scale(0.8)',
-            '-ms-transform': 'translate(-50%, -50%) scale(0.8)',
-            '-o-transform': 'translate(-50%, -50%) scale(0.8)',
-            'transform': 'translate(-50%, -50%) scale(0.8)'
-        }))
     ])
+]);
+
+export const load: AnimationMetadata =
+trigger('load', [
+    state('enter', style({
+        'opacity': '1',
+        '-webkit-transform': 'translateY(0)',
+        '-moz-transform': 'translateY(0)',
+        '-ms-transform': 'translateY(0)',
+        '-o-transform': 'translateY(0)',
+        'transform': 'translateY(0)'
+    })),
+    state('exit', style({
+        'opacity': '0',
+        '-webkit-transform': 'translateY(20%)',
+        '-moz-transform': 'translateY(20%)',
+        '-ms-transform': 'translateY(20%)',
+        '-o-transform': 'translateY(20%)',
+        'transform': 'translateY(20%)'
+    })),
+    state('void', style({
+        'opacity': '0',
+        '-webkit-transform': 'translateY(20%)',
+        '-moz-transform': 'translateY(20%)',
+        '-ms-transform': 'translateY(20%)',
+        '-o-transform': 'translateY(20%)',
+        'transform': 'translateY(20%)'
+    })),
+    transition('* => *', animate('0.3s ease-in'))
 ]);
 
 export const toast: AnimationMetadata =
 trigger('toast', [
-    state('on', style({
+    state('enter', style({
         'opacity': '1',
         '-webkit-transform': 'translate(-50%, 0)',
         '-moz-transform': 'translate(-50%, 0)',
@@ -80,7 +78,7 @@ trigger('toast', [
         '-o-transform': 'translate(-50%, 0)',
         'transform': 'translate(-50%, 0)'
     })),
-    state('init', style({
+    state('exit', style({
         'opacity': '0',
         '-webkit-transform': 'translate(-50%, 50%)',
         '-moz-transform': 'translate(-50%, 50%)',
@@ -88,7 +86,7 @@ trigger('toast', [
         '-o-transform': 'translate(-50%, 50%)',
         'transform': 'translate(-50%, 50%)'
     })),
-    state('off', style({
+    state('void', style({
         'opacity': '0',
         '-webkit-transform': 'translate(-50%, 50%)',
         '-moz-transform': 'translate(-50%, 50%)',
@@ -96,26 +94,5 @@ trigger('toast', [
         '-o-transform': 'translate(-50%, 50%)',
         'transform': 'translate(-50%, 50%)'
     })),
-    transition('init => on', [
-        style({
-            'opacity': '0',
-            '-webkit-transform': 'translate(-50%, 50%)',
-            '-moz-transform': 'translate(-50%, 50%)',
-            '-ms-transform': 'translate(-50%, 50%)',
-            '-o-transform': 'translate(-50%, 50%)',
-            'transform': 'translate(-50%, 50%)'
-        }),
-        animate('0.3s ease-in')
-    ]),
-    transition('on => off', [
-        animate('0.3s ease-out',
-        style({
-            'opacity': '0',
-            '-webkit-transform': 'translate(-50%, 50%)',
-            '-moz-transform': 'translate(-50%, 50%)',
-            '-ms-transform': 'translate(-50%, 50%)',
-            '-o-transform': 'translate(-50%, 50%)',
-            'transform': 'translate(-50%, 50%)'
-        }))
-    ])
+    transition('* => *', animate('0.3s ease-in'))
 ]);