+++ src/yup/dialog/container.ts
@@ -1,64 +1,19 @@
 import {
-    Injectable,
-    ComponentFactoryResolver,
-    ApplicationRef,
-    Injector,
-    Inject,
-    Component,
-    ViewContainerRef,
-    ViewChild,
-    EventEmitter,
-    ChangeDetectorRef,
-    ChangeDetectionStrategy
+    ComponentFactoryResolver, Injector, Component, ViewContainerRef, ViewChild,
+    EventEmitter, ChangeDetectorRef, ChangeDetectionStrategy, Inject, ComponentRef
 } from '@angular/core';
-import { animate, AnimationEvent, state, style, transition, trigger } from '@angular/animations';
 import { ComponentType, CONTAINER_DATA } from '../base/common';
 import { ContainerHostDirective } from '../base/container-host';
 import { RootRef } from '../base/root.ref';
 import { ContainerRef } from '../base/container.ref';
+import { dialog, mask } from '../animations';
 
 @Component({
     selector: 'yup-container',
     templateUrl: './container.html',
     styleUrls: ['./container.css'],
     changeDetection: ChangeDetectionStrategy.Default,
-    animations: [
-        trigger('dialog', [
-            state('enter', style({
-                'opacity': '1',
-                '-webkit-transform': 'translate(-50%, -50%) scale(1)',
-                '-moz-transform': 'translate(-50%, -50%) scale(1)',
-                '-ms-transform': 'translate(-50%, -50%) scale(1)',
-                '-o-transform': 'translate(-50%, -50%) scale(1)',
-                'transform': 'translate(-50%, -50%) scale(1)'
-            })),
-            state('void', style({
-                'opacity': '0',
-                '-webkit-transform': 'translate(-50%, -50%) scale(1.2)',
-                '-moz-transform': 'translate(-50%, -50%) scale(1.2)',
-                '-ms-transform': 'translate(-50%, -50%) scale(1.2)',
-                '-o-transform': 'translate(-50%, -50%) scale(1.2)',
-                'transform': 'translate(-50%, -50%) scale(1.2)'
-            })),
-            state('exit', style({
-                'opacity': '0',
-                '-webkit-transform': 'translate(-50%, -50%) scale(0.8)',
-                '-moz-transform': 'translate(-50%, -50%) scale(0.8)',
-                '-ms-transform': 'translate(-50%, -50%) scale(0.8)',
-                '-o-transform': 'translate(-50%, -50%) scale(0.8)',
-                'transform': 'translate(-50%, -50%) scale(0.8)'
-            })),
-            transition('* => *', [
-                animate('0.3s linear')
-            ])
-        ]),
-        trigger('mask', [
-            state('enter', style({opacity: 0.7})),
-            state('void', style({opacity: 0})),
-            state('exit', style({opacity: 0})),
-            transition('* => *', animate('0.3s linear'))
-        ])
-    ]
+    animations: [dialog, mask]
 })
 export class DialogContainerComponent {
     public animationStateChange = new EventEmitter<AnimationEvent>();
@@ -70,9 +25,9 @@ export class DialogContainerComponent {
         private viewContainerRef: ViewContainerRef,
         private componentFactoryResolver: ComponentFactoryResolver,
         private changeDetectorRef: ChangeDetectorRef,
-        @Inject(CONTAINER_DATA) public background: string
+        @Inject(CONTAINER_DATA) public data: any
     ) {}
-    public attachComponent<T>(component: ComponentType<T>, injector: Injector) {
+    public attachComponent<T>(component: ComponentType<T>, injector: Injector): ComponentRef<T> {
         return this.host.attachComponent(component, injector);
     }
     public _onAnimationDone(event: AnimationEvent) {
@@ -86,7 +41,6 @@ export class DialogContainerComponent {
     }
         /** Starts the dialog exit animation. */
     public _startExitAnimation(): void {
-        console.log('执行关闭动画');
         this._state = 'exit';
         // Mark the container for check so it can react if the
         // view container is using OnPush change detection.
@@ -94,7 +48,7 @@ export class DialogContainerComponent {
     }
     // 从遮罩关闭
     public close() {
-        if (this.background === 'mask' || this.background === 'loose') {
+        if (this.data.background === 'mask' || this.data.background === 'loose') {
             this._startExitAnimation();
         }
     }