+++ src/yup/dialog/dialog.ref.ts
@@ -21,7 +21,6 @@ export class DialogRef<T> {
         Observable.from(containerComponent.animationStateChange)
         .filter((event: AnimationEvent) => event.phaseName === 'done' && event.toState === 'exit')
         .subscribe(() => {
-            console.log('关闭动画完成');
             this.containerRef.removeComponent();
             this.afterClosed.next(this.result);
             this.afterClosed.complete();
@@ -35,7 +34,6 @@ export class DialogRef<T> {
         Observable.from(this.containerComponent.animationStateChange)
             .filter((event: AnimationEvent) => event.phaseName === 'start')
             .subscribe();
-        console.log('启动关闭');
         this.containerComponent._startExitAnimation();
     }
     public afterClose() {