+++ src/yup/toast/container.ts
@@ -1,55 +1,18 @@
 import {
-    Injectable,
-    ComponentFactoryResolver,
-    ApplicationRef,
-    Injector,
-    Inject,
-    Component,
-    ViewContainerRef,
-    ViewChild,
-    EventEmitter,
-    ChangeDetectorRef,
-    ChangeDetectionStrategy
+    ComponentFactoryResolver, Injector, Component, ViewContainerRef, ViewChild,
+    EventEmitter, ChangeDetectorRef, ChangeDetectionStrategy, ComponentRef
 } from '@angular/core';
-import { animate, AnimationEvent, state, style, transition, trigger } from '@angular/animations';
 import { ComponentType, CONTAINER_DATA } from '../base/common';
 import { ContainerHostDirective } from '../base/container-host';
 import { RootRef } from '../base/root.ref';
 import { ContainerRef } from '../base/container.ref';
+import { toast } from '../animations';
 
 @Component({
     templateUrl: './container.html',
     styleUrls: [`./container.css`],
     changeDetection: ChangeDetectionStrategy.Default,
-    animations: [
-        trigger('toast', [
-            state('enter', style({
-                'opacity': '1',
-                '-webkit-transform': 'translate(-50%, 0)',
-                '-moz-transform': 'translate(-50%, 0)',
-                '-ms-transform': 'translate(-50%, 0)',
-                '-o-transform': 'translate(-50%, 0)',
-                'transform': 'translate(-50%, 0)'
-            })),
-            state('exit', style({
-                'opacity': '0',
-                '-webkit-transform': 'translate(-50%, 50%)',
-                '-moz-transform': 'translate(-50%, 50%)',
-                '-ms-transform': 'translate(-50%, 50%)',
-                '-o-transform': 'translate(-50%, 50%)',
-                'transform': 'translate(-50%, 50%)'
-            })),
-            state('void', style({
-                'opacity': '0',
-                '-webkit-transform': 'translate(-50%, 50%)',
-                '-moz-transform': 'translate(-50%, 50%)',
-                '-ms-transform': 'translate(-50%, 50%)',
-                '-o-transform': 'translate(-50%, 50%)',
-                'transform': 'translate(-50%, 50%)'
-            })),
-            transition('* => *', animate('0.3s ease-in'))
-        ])
-    ]
+    animations: [toast]
 })
 export class ToastContainerComponent {
     public animationStateChange = new EventEmitter<AnimationEvent>();
@@ -62,7 +25,7 @@ export class ToastContainerComponent {
         private componentFactoryResolver: ComponentFactoryResolver,
         private changeDetectorRef: ChangeDetectorRef
     ) {}
-    public attachComponent<T>(component: ComponentType<T>, injector: Injector) {
+    public attachComponent<T>(component: ComponentType<T>, injector: Injector): ComponentRef<T> {
         return this.host.attachComponent(component, injector);
     }
     public _onAnimationDone(event: AnimationEvent) {
@@ -76,7 +39,6 @@ export class ToastContainerComponent {
     }
         /** Starts the dialog exit animation. */
     public _startExitAnimation(): void {
-        console.log('执行关闭动画');
         this._state = 'exit';
         // Mark the container for check so it can react if the
         // view container is using OnPush change detection.